requirements.txt
sqlalchemy == 2.0.21
FastApi ==  0.103.2
uvicorn ==  0.23.2
pydantic == 2.4.2
pyjwt == 2.8.0
argon2-cffi == 23.1.0
aiogram == 3.1.1
psycopg2-binary ==2.9.9
alembic == 1.12.0
